Add reconnect support for CH
Allow CH domains to be reconnected to after restarting libvirt. This
change enables the creation of the pidfile and domain definiton file
alongside the socket file. With these files in place (and getting
updated). The CH domain is now able to track what domains were running
when libvirt was closed. From there add the matching restore code to
allow the VM state to be read in and processed by libvirt.

Changes to hardware device management will necessitate more
development on this feature in order to track the state of different
devices in use (this is fairly different than the equivalent QEMU
code).